feat: Step1 리뷰 반영
