+++ RockPaperScissors/RockPaperScissors/Enum/GameResult.swift
@@ -1,0 +1,14 @@
+//
+//  GameResult.swift
+//  RockPaperScissors
+//
+//  Created by Janine on 2023/09/11.
+//
+
+import Foundation
+
+enum GameResult: String {
+    case draw = "비겼습니다!"
+    case win = "이겼습니다!"
+    case lose = "졌습니다!"
+}
